(Hopefully) Final UI Scaling Update
New UI design should allow for scaling of the tool based on the browser's resolution. The tool will be locked to a 16:9 ratio, so blank spaces along the left/right or top/bottom may occur due to this. Currently a black background has been set to negate the feel of this, though future updates might have set resolutions at which the tool can be run at instead of 'any'.